Moving project models to `models/` directory.
Previously, models were downloaded to the `opt/models/` directory so as not to conflict with the `/models` folder in the project. However, I think it makes more sense to include all model files in the project's `models/` directory. This also means the development container now reflects the production folder structure.